SAMZA-2211: [Cleanup] Move RunLoop to container package
